server/worker_host/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halloween-worker-host"
version = "0.1.0"
description = "Worker host that loads prop backends and bridges them to MQTT"
requires-python = ">=3.11"
dependencies = [
  "paho-mqtt>=1.6",
]

[tool.setuptools]
packages = ["worker_host"]
